+++ applications/amathea/entrypoint.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-envsubst < /etc/nginx/nginx.template > /etc/nginx/nginx.conf && exec nginx -g 'daemon off;'