+++ applications/fishing-map/src/features/dataviews/dataviews.mock.ts
@@ -3,73 +3,25 @@ import { Dataview } from '@globalfishingwatch/api-types/dist'
 export const dataviews: Dataview[] = [
   // {
   //   id: 1,
-  //   name: 'Apparent fishing effort',
+  //   name: 'Encounter cluster events',
   //   description: '',
   //   app: 'fishing-map',
   //   config: {
-  //     type: 'HEATMAP_ANIMATED',
-  //     color: '#00FFBC',
-  //     datasets: [
-  //       'public-global-fishing-effort:v20201001',
-  //       'chile-fishing:v20200331',
-  //       'indonesia-fishing:v20200320',
-  //       'panama-fishing:v20200331',
-  //       'peru-fishing:v20200324',
-  //     ],
-  //     colorRamp: 'teal',
+  //     type: 'TILE_CLUSTER',
+  //     color: '#FAE9A0',
   //   },
-  //   infoConfig: undefined,
-  //   category: DataviewCategory.Activity,
+  //   category: DataviewCategory.Events,
   //   datasetsConfig: [
   //     {
-  //       params: [
+  //       query: [
   //         {
-  //           id: 'type',
-  //           value: 'heatmap',
+  //           id: 'types',
+  //           value: ['encounter'],
   //         },
   //       ],
-  //       endpoint: '4wings-tiles',
-  //       datasetId: 'public-global-fishing-effort:v20201001',
-  //     },
-  //     {
-  //       params: [
-  //         {
-  //           id: 'type',
-  //           value: 'heatmap',
-  //         },
-  //       ],
-  //       endpoint: '4wings-tiles',
-  //       datasetId: 'chile-fishing:v20200331',
-  //     },
-  //     {
-  //       params: [
-  //         {
-  //           id: 'type',
-  //           value: 'heatmap',
-  //         },
-  //       ],
-  //       endpoint: '4wings-tiles',
-  //       datasetId: 'indonesia-fishing:v20200320',
-  //     },
-  //     {
-  //       params: [
-  //         {
-  //           id: 'type',
-  //           value: 'heatmap',
-  //         },
-  //       ],
-  //       endpoint: '4wings-tiles',
-  //       datasetId: 'peru-fishing:v20200324',
-  //     },
-  //     {
-  //       params: [
-  //         {
-  //           id: 'type',
-  //           value: 'heatmap',
-  //         },
-  //       ],
-  //       endpoint: '4wings-tiles',
-  //       datasetId: 'panama-fishing:v20200331',
+  //       params: [],
+  //       endpoint: 'carriers-events-cluster-tiles',
+  //       datasetId: 'carriers-clusterbuster',
   //     },
   //   ],
   // },