+++ applications/fishing-map/src/features/dataviews/dataviews.utils.ts
@@ -1,4 +1,4 @@
-import { DataviewCategory, DataviewInstance } from '@globalfishingwatch/api-types'
+import { DataviewCategory, DataviewInstance, EndpointId } from '@globalfishingwatch/api-types'
 import {
   TrackColorBarOptions,
   HeatmapColorBarOptions,
@@ -11,7 +11,9 @@ import {
   DEFAULT_VESSEL_DATAVIEW_ID,
 } from 'data/workspaces'
 
-export const DATAVIEW_INSTANCE_PREFIX = 'vessel-'
+// used in workspaces with encounter events layers
+export const ENCOUNTER_EVENTS_SOURCE_ID = 'encounter-events'
+export const VESSEL_LAYER_PREFIX = 'vessel-'
 export const ENVIRONMENTAL_LAYER_PREFIX = 'environment-'
 export const CONTEXT_LAYER_PREFIX = 'context-'
 
@@ -27,18 +29,19 @@ export const getVesselDataviewInstance = (
     {
       datasetId: trackDatasetId,
       params: [{ id: 'vesselId', value: vessel.id }],
-      endpoint: 'carriers-tracks',
+      endpoint: EndpointId.Tracks,
     },
     {
       datasetId: infoDatasetId,
       params: [{ id: 'vesselId', value: vessel.id }],
-      endpoint: 'carriers-vessel',
+      endpoint: EndpointId.Vessel,
     },
   ]
   const vesselDataviewInstance = {
-    id: `${DATAVIEW_INSTANCE_PREFIX}${vessel.id}`,
+    id: `${VESSEL_LAYER_PREFIX}${vessel.id}`,
     dataviewId: DEFAULT_VESSEL_DATAVIEW_ID,
     config: {
+      // TODO pick a not used color
       color: TrackColorBarOptions[Math.floor(Math.random() * TrackColorBarOptions.length)].value,
     },
     datasetsConfig,
@@ -80,7 +83,7 @@ export const getEnvironmentDataviewInstance = (
       {
         datasetId,
         params: [],
-        endpoint: 'user-context-tiles',
+        endpoint: EndpointId.UserContextTiles,
       },
     ],
   }