+++ applications/vessel-history/src/features/profile/components/Info.tsx
@@ -1,12 +1,11 @@
 import React, { Fragment } from 'react'
 import { Button } from '@globalfishingwatch/ui-components'
-// eslint-disable-next-line import/order
-import { VesselInfo } from 'classes/vessel.class'
+import { VesselWithHistory } from 'types'
 import styles from './Info.module.css'
 import InfoField from './InfoField'
 
 interface InfoProps {
-  vessel: VesselInfo | null
+  vessel: VesselWithHistory | null
   lastPosition: any
   lastPortVisit: any
 }
@@ -20,71 +19,76 @@ const Info: React.FC<InfoProps> = (props): React.ReactElement => {
         {vessel && (
           <div className={styles.identifiers}>
             <InfoField
-              vesselName={vessel.getName().value?.data ?? ''}
+              vesselName={vessel.shipname ?? ''}
               label={'Name'}
-              field={vessel.getName()}
+              value={vessel.shipname}
+              valuesHistory={vessel.history.shipname.byDate}
             ></InfoField>
             <InfoField
-              vesselName={vessel.getName().value?.data ?? ''}
+              vesselName={vessel.shipname ?? ''}
               label={'Type'}
-              field={vessel.getType()}
+              value={vessel.type ?? ''}
             ></InfoField>
             <InfoField
-              vesselName={vessel.getName().value?.data ?? ''}
+              vesselName={vessel.shipname ?? ''}
               label={'FLAG'}
-              field={vessel.getFlag()}
+              value={vessel.flag}
+              valuesHistory={vessel.history.flag.byDate}
             ></InfoField>
             <InfoField
-              vesselName={vessel.getName().value?.data ?? ''}
+              vesselName={vessel.shipname ?? ''}
               label={'MMSI'}
-              field={vessel.getMMSI()}
+              value={vessel.mmsi ?? ''}
+              valuesHistory={vessel.history.mmsi.byDate}
             ></InfoField>
             <InfoField
-              vesselName={vessel.getName().value?.data ?? ''}
+              vesselName={vessel.shipname ?? ''}
               label={'CALLSIGN'}
-              field={vessel.getCallsign()}
+              value={vessel.callsign ?? ''}
+              valuesHistory={vessel.history.callsign.byDate}
             ></InfoField>
             <InfoField
-              vesselName={vessel.getName().value?.data ?? ''}
+              vesselName={vessel.shipname ?? ''}
               label={'GEAR TYPE'}
-              field={vessel.getGearType()}
+              value={vessel.gearType ?? ''}
             ></InfoField>
             <InfoField
-              vesselName={vessel.getName().value?.data ?? ''}
+              vesselName={vessel.shipname ?? ''}
               label={'LENGTH'}
-              field={vessel.getLength()}
+              value={vessel.length ?? ''}
             ></InfoField>
             <InfoField
-              vesselName={vessel.getName().value?.data ?? ''}
+              vesselName={vessel.shipname ?? ''}
               label={'GROSS TONNAGE'}
-              field={vessel.getGrossTonnage()}
+              value={vessel.grossTonnage ?? ''}
             ></InfoField>
             <InfoField
-              vesselName={vessel.getName().value?.data ?? ''}
+              vesselName={vessel.shipname ?? ''}
               label={'DEPTH'}
-              field={vessel.getDepth()}
+              value={vessel.depth ?? ''}
             ></InfoField>
             <div className={styles.identifierField}>
               <label>AUTHORIZATIONS</label>
-              {vessel.getAuthorisations().map((auth) => (
+              {vessel.authorizations?.map((auth) => (
                 <p key={auth}>{auth}</p>
               ))}
-              {!vessel.getAuthorisations().length && <p>No authorizations found</p>}
+              {!vessel.authorizations?.length && <p>No authorizations found</p>}
             </div>
             <InfoField
-              vesselName={vessel.getName().value?.data ?? ''}
+              vesselName={vessel.shipname ?? ''}
               label={'BUILT'}
-              field={vessel.getBuiltYear()}
+              value={vessel.builtYear ?? ''}
             ></InfoField>
             <InfoField
-              vesselName={vessel.getName().value?.data ?? ''}
+              vesselName={vessel.shipname ?? ''}
               label={'OWNER'}
-              field={vessel.getOwner()}
+              value={vessel.owner ?? ''}
+              valuesHistory={vessel.history.owner.byDate}
             ></InfoField>
             <InfoField
-              vesselName={vessel.getName().value?.data ?? ''}
+              vesselName={vessel.shipname ?? ''}
               label={'OPERATOR'}
-              field={vessel.getOperator()}
+              value={vessel.operator ?? ''}
             ></InfoField>
           </div>
         )}