+++ packages/api-types/src/vessel.ts
@@ -9,9 +9,20 @@ export interface Vessel {
   callsign?: string
   fleet?: string
   origin?: string
+  type?: string
+  gearType?: string
+  length?: string
+  depth?: string
+  grossTonnage?: string
+  owner?: string
+  operator?: string
+  builtYear?: string
+  authorizations?: string[]
+  registeredGearType?: string
 }
 
 export interface VesselSearch extends Vessel {
   dataset: string
   source: string // Label of the dataset
+  vesselMatchId: string
 }