+++ packages/react-hooks/src/use-map-interaction/index.ts
@@ -1,16 +1,8 @@
 import type { Geometry } from 'geojson'
-import { Dataset } from '@globalfishingwatch/api-types'
 import { ContextLayerType } from '@globalfishingwatch/layer-composer/dist/generators/types'
 import { PointerEvent } from '@globalfishingwatch/react-map-gl'
 
 export { useMapHover, useMapClick, useFeatureState } from './use-map-interaction'
-
-export type ExtendedFeatureVessel = {
-  id: string
-  hours: number
-  dataset: Dataset
-  [key: string]: any
-}
 
 export type ExtendedFeature = {
   properties: Record<string, any>
@@ -34,8 +26,6 @@ export type ExtendedFeature = {
     row: number
   }
   generatorContextLayer?: ContextLayerType | null
-  vessels?: ExtendedFeatureVessel[]
-  dataset?: Dataset
   geometry?: Geometry
 }